Prepare data for DEG anlysis

S3 general function DEG_prepareData(), prepare the data that LZ diff analysis need,
return a list(eset=xx, group=xx, f_mark="xx")

Usage

DEG_prepareData(
  object,
  eset_file,
  eset.islog,
  id_dot,
  col.by,
  col.del = NULL,
  auto.del.character = T,
  annot_trans,
  group_file,
  f_mark = "diff",
  is.human = T,
  orgDb = "org.Mm.eg.db",
  fromType = "ENSEMBL",
  toType = c("SYMBOL", "UNIPROT"),
  oop = FALSE,
  oop.group.suffix1,
  oop.group.suffix2,
  oop.group.group1,
  oop.group.group2,
  ...
)

# S3 method for default
DEG_prepareData(
  ...,
  eset_file = "gene_count.csv",
  eset.islog = F,
  id_dot = F,
  col.by = "ID",
  col.del = NULL,
  auto.del.character = T,
  annot_trans = T,
  group_file,
  f_mark = "diff",
  is.human = T,
  orgDb = "org.Mm.eg.db",
  fromType = "ENSEMBL",
  toType = c("SYMBOL", "UNIPROT"),
  oop = FALSE,
  oop.group.suffix1 = "01A",
  oop.group.suffix2 = "11A",
  oop.group.group1 = "Tumor",
  oop.group.group2 = "Normal"
)

# S3 method for XENA
DEG_prepareData(
  object,
  ...,
  eset.islog = T,
  id_dot = T,
  col.by = "Ensembl_ID",
  col.del = NULL,
  auto.del.character = T,
  annot_trans = T,
  group_file,
  f_mark = "diff",
  is.human = T,
  orgDb = "org.Mm.eg.db",
  fromType = "ENSEMBL",
  toType = c("SYMBOL", "UNIPROT"),
  oop = T,
  oop.group.suffix1 = "01A",
  oop.group.suffix2 = "11A",
  oop.group.group1 = "Tumor",
  oop.group.group2 = "Normal"
)

# S3 method for DEeset
DEG_prepareData(object, ..., f_mark = "DE")

Arguments

object

any R object

eset_file

character, the exps data file name, default: "gene_count.csv"

eset.islog

logical, if the eset have been log2, default is F, meaning the eset data is not log2

id_dot

logical, if the id column is ensemble id with dot, ex.ESEMxxxx.3

col.by

character, delete the duplicate by this column

col.del

character, manually delete some column

auto.del.character

logical, if auto delete the column those class is character

annot_trans

logical, do the eset need be annot to symbol? if the datafame's first coloumn is not offical symbol, you can set to TRUE.

group_file

the group data file name, such as: "group.csv", if the param group_file is missing and oop = T this function will use DEeset$updateGroup() method

f_mark

file mark(optional), default: diff

is.human

logical, default is True, if it is True, the param orgDB,fromTyp and toType will be ignore

orgDb

character, annotation db, ex. "org.Mm.eg.db"

fromType

character, eset id column type

toType

character, the annot result

oop

logical, if use oop flow

oop.group.suffix1

character, default 01A, eset's name end with 01A will use as group Tumor, DEeset$updateGroup()

oop.group.suffix2

character, default 01A, eset's name end with 11A will use as group Normal, DEeset$updateGroup()

oop.group.group1

character, default Tumor, eset's name end with 01A will use as group Tumor, DEeset$updateGroup()

oop.group.group2

character, default Normal, eset's name end with 01A will use as group Normal, DEeset$updateGroup()

...

other param

Author

Jiang

Examples

# read from XENABD object, oop  ============
# DEeset <- DEG_prepareData(XENADB,
#                           id_dot = T, col.by = "Ensembl_ID",
#                           col.del=NULL, auto.del.character=T,
#                           annot_trans=T, f_mark="diff",
#                           oop = T,
#                           oop.group.suffix1 = "01A",
#                           oop.group.suffix2 = "11A",
#                           oop.group.endT = "Tumor",
#                           oop.group.endF = "Normal")
#
# read from local file, oop ============
# test <- DEG_prepareData(eset_file="D:/TCGA.counts.tsv.gz",
#                         id_dot = T, col.by = "Ensembl_ID",
#                         col.del=NULL, auto.del.character=T,
#                         annot_trans=T, f_mark="diff",
#                         oop = T,
#                         oop.group.suffix1 = "01A",
#                         oop.group.suffix2 = "11A",
#                         oop.group.endT = "Tumor",
#                         oop.group.endF = "Normal")
#
# read frome local file(eset, group), no oop ============
# 仅当oop = F, 才会进入非oop流程, 此时需要指定group_flie参数才行。默认情况为F
# glist <- DEG_prepareData(eset_file = "data/eset.csv", #表达数据的相对路径
#                          group_file = "data/group.csv", #分组文件的相对路径
#                          id_dot = F,  # ESEM是否有点，有点设为T
#                          col.by = "ID",  # 基因名列的列名
#                          annot_trans = F, # 是否要注释，如果是EsembleID就需要设置为T
#                          f_mark = mark)
#' #' # exprset.group = list(eset, group, f_mark)
# ** 如果bacth=T，则group中第二列为batch列。
# ** 矩阵数据格式(数值型，整型)
# gene  | row1 | row2 | row3  | row4
# gene1 |  34  |  23  |  56   |  23
# gene2 |  35  |  23  |  12   |  23
# gene3 |  12  |  78  |  78   |  78
# ** 分组数据格式：需要组的行名 包含于 表达谱的列名 rownames(group) %in% colname(eset)
# Sample   |   Type    | BATCH
# rowname1 |   tumor   |  1
# rowname2 |   tumor   |  1
# rowname3 |   normal  |  2
# rowname4 |   normal  |  2
# if is.human = F, please set the orgDb and fromType correctly.