Eeset t.test or wilcox.test

Eeset t.test or wilcox.test

Usage

DEG_tw.test(
  eset,
  group,
  method = "wilcox.test",
  pval = 0.05,
  fdr = 0.1,
  logfc = log2(2)
)

Arguments

eset

data.frame, the exprs data, the data must not be log2 transformed

group

data.frame, the pheno information, the first coloumn is the type information and should be factor.

method

character, the method of test, default is wilcox.test, alternative is t.test

pval

number, the pvalue cutoff

fdr

number, the fdr cutoff

logfc

number, the log2(fc) value cutoff

Value

list(resdf = res, deg = deg)

Author

Jiang

Examples

if (FALSE) {
eset <- data.table::fread("./protein_count_pn.csv", data.table = F)
eset <- eset %>% tibble::column_to_rownames(var = "Protein")

group <- read.csv("./group.csv", row.names = 1)
group$Type <- as.factor(group$Type)

diff <- DEG_tw.test(eset = eset, group = group)
}