GO all barplot

GO all barplot

Usage

DEGp_GOALL_bar(
  GOlist.df,
  resultdir,
  filemark,
  title = "GO",
  pcompress = 2,
  x.which = "pvalue",
  pic.save = F,
  width = 10,
  height = 10
)

Arguments

GOlist.df

dataframe, GO result, should contain column "ONTOLOGY", "Description", "pvalue", "Count"

resultdir

character, the picture outdir

filemark

character, the picture filename mark

title

character, the picture title

pcompress

number, create a blank area at the top of the picture

x.which

character, x axis which to use, "pvalue" or "Count"

pic.save

logical, save picture or not. if F, the param width and height will not work

width

number, picture width

height

number, picture height

Value

ggplot2 picture

Author

Jiang