S3 general function DEG_prepareData()

S3 general function DEG_prepareData(), prepare the data that LZ diff analysis need,
return a list(eset=xx, group=xx, f_mark="xx")
DEG_prepareData.default default method
DEG_prepareData.DEeset DEeset method

Usage

DEG_prepareData(...)

# S3 method for default
DEG_prepareData(
  eset_file = "gene_count.csv",
  id_dot = F,
  col.by = "ID",
  col.del = NULL,
  auto.del.character = T,
  group_file = "group.csv",
  annot_trans = T,
  f_mark = "diff",
  is.human = T,
  orgDb = "org.Mm.eg.db",
  fromType = "ENSEMBL",
  toType = c("SYMBOL", "UNIPROT")
)

# S3 method for DEeset
DEG_prepareData(obj, f_mark = "DE")

# S3 method for xena
DEG_prepareData(eset, group)

Arguments

eset_file

character, the exps data file name, default: "gene_count.csv"

id_dot

logical, if the id column is ensemble id with dot, ex.ESEMxxxx.3

col.by

character, delete the duplicate by this column

col.del

character, manually delete some column

auto.del.character

logical, if auto delete the column those class is character

group_file

the group data file name, default: "group.csv"

annot_trans

logical, do the eset need be annot to symbol? if the datafame's first coloumn is not offical symbol, you can set to TRUE.

f_mark

character a analysis mark(optional), default: "DE"

is.human

logical, default is True, if it is True, the param orgDB,fromTyp and toType will be ignore

orgDb

character, annotation db, ex. "org.Mm.eg.db"

fromType

character, eset id column type

toType

character, the annot result

obj

obeject the object DEeset

Value

list

list

list

Examples

#
#' # exprset.group = list(eset, group, f_mark)
# ** 如果bacth=T，则group中第二列为batch列。
# ** 矩阵数据格式(数值型，整型)
# gene  | row1 | row2 | row3  | row4
# gene1 |  34  |  23  |  56   |  23
# gene2 |  35  |  23  |  12   |  23
# gene3 |  12  |  78  |  78   |  78
# ** 分组数据格式：需要组的行名 包含于 表达谱的列名 rownames(group) %in% colname(eset)
# Sample   |   Type    | BATCH
# rowname1 |   tumor   |  1
# rowname2 |   tumor   |  1
# rowname3 |   normal  |  2
# rowname4 |   normal  |  2
# if is.human = F, please set the orgDb and fromType correctly.